pyproject.toml
[tool.poetry]
name = "level-learner"
version = "0.1.0"
description = ""
authors = ["AJ laptop <talkingtoaj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-fasthtml = "*"
pydantic = "*"
markdown = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
